+++ src/main/java/finalmission/domain/Member.java
@@ -1,4 +1,4 @@
-package finalmission.entity;
+package finalmission.domain;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
