+++ src/main/java/finalmission/domain/Reservation.java
@@ -1,4 +1,4 @@
-package finalmission.entity;
+package finalmission.domain;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
