+++ src/main/java/finalmission/repository/MemberRepository.java
@@ -1,6 +1,6 @@
 package finalmission.repository;
 
-import finalmission.entity.Member;
+import finalmission.domain.Member;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
