+++ src/main/java/finalmission/repository/MusicalRepository.java
@@ -1,6 +1,6 @@
 package finalmission.repository;
 
-import finalmission.entity.Musical;
+import finalmission.domain.Musical;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MusicalRepository extends JpaRepository<Musical, Long> {
