+++ src/main/java/finalmission/repository/SeatRepository.java
@@ -1,6 +1,6 @@
 package finalmission.repository;
 
-import finalmission.entity.Seat;
+import finalmission.domain.Seat;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SeatRepository extends JpaRepository<Seat, Long> {
