+++ src/main/java/finalmission/service/ReservationService.java
@@ -6,11 +6,11 @@ import finalmission.dto.LoginMemberInfo;
 import finalmission.dto.ReservationFullRequest;
 import finalmission.dto.ReservationFullResponse;
 import finalmission.dto.ReservationSimpleResponse;
-import finalmission.entity.Member;
-import finalmission.entity.Musical;
-import finalmission.entity.MusicalTime;
-import finalmission.entity.Reservation;
-import finalmission.entity.Seat;
+import finalmission.domain.Member;
+import finalmission.domain.Musical;
+import finalmission.domain.MusicalTime;
+import finalmission.domain.Reservation;
+import finalmission.domain.Seat;
 import finalmission.exception.UnauthorizedException;
 import finalmission.repository.MemberRepository;
 import finalmission.repository.MusicalRepository;
