+++ src/test/java/finalmission/service/ReservationServiceTest.java
@@ -12,10 +12,10 @@ import finalmission.client.dto.HolidaysResponse.HolidayItems;
 import finalmission.client.dto.HolidaysResponse.HolidayResponse;
 import finalmission.dto.LoginMemberInfo;
 import finalmission.dto.ReservationFullRequest;
-import finalmission.entity.Member;
-import finalmission.entity.Musical;
-import finalmission.entity.Seat;
-import finalmission.entity.SeatGrade;
+import finalmission.domain.Member;
+import finalmission.domain.Musical;
+import finalmission.domain.Seat;
+import finalmission.domain.SeatGrade;
 import finalmission.repository.MemberRepository;
 import finalmission.repository.MusicalRepository;
 import finalmission.repository.ReservationRepository;
